# OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: RiskModels API
  version: 3.0.0-agent
  description: >
    Precision equity risk intelligence platform with AI agent integration.

    Institutional-grade equity risk analysis for AI agents and quantitative applications.
    Provides daily factor hedge ratios, explained-risk decompositions, and portfolio-level
    risk attribution for ~3,000 US equities.


    **New in v3.0.0-agent:**

    - OAuth2 client credentials flow for machine-to-machine auth

    - Plaid Investments integration for live portfolio sync

    - MCP (Model Context Protocol) server for AI agent tools

    - Compliance manifests for AI marketplace integration

    - Per-key rate limiting and scoped access control


    **Data**: Universe `uni_mc_3000` (top ~3,000 US stocks by market cap), date range
    2006-01-04 to present, updated daily end-of-day. Backend: Zarr v2 on Google Cloud Storage,
    computed by the ERM3 regression system (Huber/Ridge). Supabase tables `erm3_betas` and
    `erm3_rankings` provide factor betas and ticker rankings for direct DB and API use.


    **Authentication Methods:**

    1. Session cookies (web app)

    2. API Key/Secret via Bearer token (external integrations)

    3. OAuth2 client credentials (AI agents, npm package)


    **Security Features:**

    - GCP KMS envelope encryption for portfolio data

    - Cryptographic shredding on key deletion

    - HMAC-SHA256 webhook verification

    - Row-level security (RLS) via Supabase


...

...

...

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: rm_agent_{environment}_{random}_{checksum}
      description: >
        API key in format `rm_agent_{live|test}_{random}_{checksum}` or `rm_user_{random}_{checksum}`.
        Obtain via `POST /api/auth/provision` or from riskmodels.app/settings → API Keys.
        Include as: `Authorization: Bearer rm_agent_live_...`

    OAuth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://riskmodels.app/api/auth/token
          scopes:
            ticker-returns: Access ticker returns and historical data
            risk-decomposition: Access L3 risk decomposition
            batch-analysis: Perform portfolio batch analysis
            factor-correlation: Correlate stocks with macro factors (VIX, Bitcoin, Gold, etc.)
            chat-risk-analyst: Use AI risk analyst
            plaid:holdings: Access Plaid-synced portfolio holdings
            "*": Full API access
...
  schemas:
    AgentMeta:
      type: object
      description: Per-request billing and telemetry metadata appended to metered responses.
      properties:
        cost_usd:
          type: number
          format: float
          description: Cost deducted from prepaid balance for this request. 0 if served from cache.
          example: 0.005
        latency_ms:
          type: integer
          description: Server processing time in milliseconds.
          example: 145
        request_id:
          type: string
          description: Unique request identifier for support lookups.
          example: req_abc123xyz
        confidence:
          type: object
          properties:
            overall:
              type: number
              format: float
              minimum: 0
              maximum: 1
              example: 0.98
            factors:
              type: object
              additionalProperties:
                type: number
        data_freshness:
          type: string
          format: date-time
          description: ISO timestamp of the data's as-of datetime.
          example: "2026-02-21T10:30:00Z"
        billing_code:
          type: string
          description: Internal billing classification for this request type.
          example: ticker_returns_v2
        cache_status:
          type: string
          enum: [HIT, MISS, BYPASS]
          description: Whether this response was served from cache.
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code.
          example: TICKER_NOT_FOUND
        message:
          type: string
          description: Human-readable error description with suggested action.
          example: "Ticker 'XYZABC' not found in universe 'uni_mc_3000'. Check ticker spelling or try a different universe."
        code:
          type: integer
          description: HTTP status code.
          example: 404
        details:
          type: object
          properties:
            field:
              type: string
              example: ticker
            received:
              type: string
              example: XYZABC
            universe:
              type: string
              example: uni_mc_3000
            suggestion:
              type: string
              example: "Check for typos or use ticker search endpoint"
...
    PlaidHolding:
      type: object
      properties:
        account_id:
          type: string
        security_id:
          type: string
        institution_value:
          type: number
          description: Current value in USD
        quantity:
          type: number
        ticker:
          type: string
          description: Normalized ticker symbol
        name:
          type: string
          description: Security name
        risk_metrics:
          $ref: "#/components/schemas/RiskMetrics"
    RiskMetrics:
      type: object
      description: Optional risk enrichment on Plaid holdings; names may match batch `full_metrics` or V3 `metrics` keys depending on pipeline version.
      properties:
        annualized_volatility:
          type: number
          nullable: true
        volatility:
          type: number
          nullable: true
        sharpe_ratio:
          type: number
          nullable: true
        l3_mkt_hr:
          type: number
          nullable: true
        l3_sec_hr:
          type: number
          nullable: true
        l3_sub_hr:
          type: number
          nullable: true
        l3_res_er:
          type: number
          nullable: true
        l3_market_hr:
          type: number
          nullable: true
        l3_sector_hr:
          type: number
          nullable: true
        l3_subsector_hr:
          type: number
          nullable: true
        l3_residual_er:
          type: number
          nullable: true
...
    InsufficientBalance:
      type: object
      description: Returned when the account prepaid balance is exhausted (HTTP 402).
      properties:
        error:
          type: string
          enum: [INSUFFICIENT_BALANCE]
          example: INSUFFICIENT_BALANCE
        message:
          type: string
          example: "Insufficient balance. Current balance: $0.00. Top up at https://riskmodels.app/billing"
        balance_usd:
          type: number
          format: float
          example: 0.0
        top_up_url:
          type: string
          format: uri
          example: https://riskmodels.app/billing
    RateLimitExceeded:
      type: object
      description: Returned when the per-minute rate limit is exceeded (HTTP 429).
      properties:
        error:
          type: string
          enum: [RATE_LIMIT_EXCEEDED]
          example: RATE_LIMIT_EXCEEDED
        message:
          type: string
          example: "Rate limit exceeded. Retry after 2026-03-08T12:34:56Z"
        retry_after:
          type: string
          format: date-time
          description: ISO timestamp after which the client may retry.
          example: "2026-03-08T12:34:56Z"
        limit:
          type: integer
          description: Requests allowed per minute for this key.
          example: 60
security:
  - BearerAuth: []
paths:
...
  /plaid/link-token:
    post:
      summary: Create Plaid Link token (Investments)
      description: >
        One-time `link_token` for Plaid Link in the browser. Requires an authenticated user.
        After Link succeeds, exchange the `public_token` via `POST /plaid/exchange-public-token`.
        **No per-request charge.**
      operationId: createPlaidLinkToken
      tags: [Plaid Integration]
      security:
        - BearerAuth: []
        - OAuth2ClientCredentials: [plaid:holdings]
      responses:
        "200":
          description: Link token for Plaid Link SDK.
          content:
            application/json:
              schema:
                type: object
                properties:
                  link_token:
                    type: string
                  expiration:
                    type: string
                    description: RFC-3339 datetime when the link_token expires.
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Plaid or configuration error.
  /plaid/exchange-public-token:
    post:
      summary: Exchange Plaid public token for stored item
      description: >
        Exchanges the short-lived `public_token` from Plaid Link for an access token, encrypted
        and stored server-side. Repeat connections create/update one row per Plaid `item_id`.
        **No per-request charge.**
      operationId: exchangePlaidPublicToken
      tags: [Plaid Integration]
      security:
        - BearerAuth: []
        - OAuth2ClientCredentials: [plaid:holdings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_token]
              properties:
                public_token:
                  type: string
                  description: From Plaid Link onSuccess.
      responses:
        "200":
          description: Item stored.
          content:
            application/json:
              schema:
                type: object
                properties:
                  item_id:
                    type: string
                  request_id:
                    type: string
        "400":
          description: Missing or invalid body.
        "401":
          description: Unauthorized.
        "500":
          description: Plaid or storage error.
  /plaid/holdings:
    get:
      summary: Fetch Plaid-synced portfolio holdings
      description: >
        Returns enriched holdings from connected Plaid accounts with risk metrics.


        **Features:**

        - Automatic ticker resolution and normalization

        - Risk enrichment with factor exposures

        - Real-time portfolio valuation

        - Multi-account aggregation


        **Authentication:** Required (API Key or OAuth2 with plaid:holdings scope)

        **Billing:** $0.01 per request

        **Rate Limit:** 60 requests/minute
      operationId: getPlaidHoldings
      tags:
        - Plaid Integration
      security:
        - BearerAuth: []
        - OAuth2ClientCredentials: [plaid:holdings]
      responses:
        "200":
          description: Enriched holdings with risk metrics
          headers:
            X-API-Cost-USD:
              schema: { type: string }
            X-RateLimit-Limit:
              schema: { type: integer }
            X-RateLimit-Remaining:
              schema: { type: integer }
            X-RateLimit-Reset:
              schema: { type: integer }
          content:
            application/json:
              schema:
                type: object
                properties:
                  holdings:
                    type: array
                    items:
                      $ref: "#/components/schemas/PlaidHolding"
                  accounts:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                    description: Plaid account objects from Investments Holdings (per connected item).
                  securities:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                    description: Plaid security objects referenced by holdings.
                  connections_count:
                    type: integer
                    description: Number of connected Plaid items for this user.
                  summary:
                    type: object
                    properties:
                      total_value:
                        type: number
                      account_count:
                        type: integer
                      position_count:
                        type: integer
                  _metadata:
                    type: object
                    additionalProperties: true
                  _agent:
                    $ref: "#/components/schemas/AgentMeta"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: API key missing `plaid:holdings` scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Insufficient balance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientBalance"
        "429":
          description: Rate limit exceeded.
          headers:
            Retry-After:
              schema: { type: integer }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitExceeded"
        "503":
          description: Plaid or token encryption not configured.
...